line_coordinates/requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cryptography==37.0.1
cycler==0.11.0
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.4
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
munkres==1.1.4
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
pip==22.2.2
ply==3.11
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.7.3
setuptools==63.4.1
sip==6.6.2
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.0
